refactor: Refactor exec_heredoc.c
