+++ Makefile
@@ -6,7 +6,7 @@
 #    By: nkim <nkim@student.42seoul.kr>             +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/05/04 14:54:52 by hannkim           #+#    #+#              #
-#    Updated: 2022/06/30 16:29:17 by nkim             ###   ########.fr        #
+#    Updated: 2022/06/30 17:38:37 by nkim             ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -77,7 +77,7 @@ SRC_TEST		= test_token.c test_ast.c test_builtin.c test_general.c
 
 SRC_EXEC_DIR	= exec/
 SRC_EXEC		= exec_command_line.c \
-					exec_ast.c exec_general.c exec_builtin.c \
+					exec_heredoc.c exec_general.c exec_builtin.c \
 					exec_subshell.c exec_single_command.c \
 					exec_simple_command.c exec_io_redirect.c
